Компонента для призываеомго меча и щит для культо-меча

+++ code/game/gamemodes/cult/talisman.dm
@@ -21,10 +21,9 @@
 		return
 
 	user.take_overall_damage(rand(5, 8), rand(1, 5))
-	rite?.perform_rite(user, src)
-
-	if(disposable)
-		qdel(src)
+	if(rite?.perform_rite(user, src))
+		if(disposable)
+			qdel(src)
 
 /obj/item/weapon/paper/talisman/chaplain/examine(mob/user)
 	..()
